Apply 15 edits across 2 files
